Reworking AWS mocks and fixing test
