+++ app/aws/aws_mocks.py
@@ -1,6 +1,78 @@
 import json
 
 
+def sns_s3_callback(filename, message_id="some-message-id"):
+    message_contents = '{"Records":[{"s3":{"object":{"key":"%s"}}}]}' % (filename)  # noqa
+    return json.dumps(
+        {
+            "Type": "Notification",
+            "MessageId": message_id,
+            "Message": message_contents,
+        }
+    )
+
+
+def ses_notification_callback(reference):
+    ses_message_body = {
+        "delivery": {
+            "processingTimeMillis": 2003,
+            "recipients": ["success@simulator.amazonses.com"],
+            "remoteMtaIp": "123.123.123.123",
+            "reportingMTA": "a7-32.smtp-out.eu-west-1.amazonses.com",
+            "smtpResponse": "250 2.6.0 Message received",
+            "timestamp": "2017-11-17T12:14:03.646Z",
+        },
+        "mail": {
+            "commonHeaders": {
+                "from": ["TEST <TEST@notify.works>"],
+                "subject": "lambda test",
+                "to": ["success@simulator.amazonses.com"],
+            },
+            "destination": ["success@simulator.amazonses.com"],
+            "headers": [
+                {"name": "From", "value": "TEST <TEST@notify.works>"},
+                {"name": "To", "value": "success@simulator.amazonses.com"},
+                {"name": "Subject", "value": "lambda test"},
+                {"name": "MIME-Version", "value": "1.0"},
+                {
+                    "name": "Content-Type",
+                    "value": 'multipart/alternative; boundary="----=_Part_617203_1627511946.1510920841645"',
+                },
+            ],
+            "headersTruncated": False,
+            "messageId": reference,
+            "sendingAccountId": "12341234",
+            "source": '"TEST" <TEST@notify.works>',
+            "sourceArn": "arn:aws:ses:eu-west-1:12341234:identity/notify.works",
+            "sourceIp": "0.0.0.1",
+            "timestamp": "2017-11-17T12:14:01.643Z",
+        },
+        "notificationType": "Delivery",
+    }
+
+    return {
+        "Type": "Notification",
+        "MessageId": "8e83c020-1234-1234-1234-92a8ee9baa0a",
+        "TopicArn": "arn:aws:sns:eu-west-1:12341234:ses_notifications",
+        "Subject": None,
+        "Message": json.dumps(ses_message_body),
+        "Timestamp": "2017-11-17T12:14:03.710Z",
+        "SignatureVersion": "1",
+        "Signature": "[REDACTED]",
+        "SigningCertUrl": "https://sns.eu-west-1.amazonaws.com/SimpleNotificationService-[REDACTED].pem",
+        "UnsubscribeUrl": "https://sns.eu-west-1.amazonaws.com/?Action=Unsubscribe&SubscriptionArn=[REACTED]",
+        "MessageAttributes": {},
+    }
+
+
+def ses_hard_bounce_callback(reference, bounce_subtype=None):
+    return _ses_bounce_callback(reference, "Permanent", bounce_subtype)
+
+
+def ses_soft_bounce_callback(reference, bounce_subtype=None):
+    return _ses_bounce_callback(reference, "Transient", bounce_subtype)
+
+
 def ses_complaint_callback_malformed_message_id():
     return {
         "Signature": "bb",
@@ -53,31 +125,6 @@ def ses_complaint_callback():
         "Message": '{"notificationType":"Complaint","complaint":{"complaintFeedbackType": "abuse", "complainedRecipients":[{"emailAddress":"recipient1@example.com"}],"timestamp":"2018-06-05T13:59:58.000Z","feedbackId":"ses_feedback_id"},"mail":{"timestamp":"2018-06-05T14:00:15.950Z","source":"\\"Some Service\\" <someservicenotifications.service.gov.uk>","sourceArn":"arn:identity/notifications.service.gov.uk","sourceIp":"52.208.24.161","sendingAccountId":"888450439860","messageId":"ref1","destination":["recipient1@example.com"]}}',  # noqa
         "SigningCertUrl": "https://sns.pem",
     }
-
-
-def ses_notification_callback():
-    return (
-        '{\n  "Type" : "Notification",\n  "MessageId" : "ref1",'
-        '\n  "TopicArn" : "arn:aws:sns:eu-west-1:123456789012:testing",'
-        '\n  "Message" : "{\\"notificationType\\":\\"Delivery\\",'
-        '\\"mail\\":{\\"timestamp\\":\\"2016-03-14T12:35:25.909Z\\",'
-        '\\"source\\":\\"test@smtp_user\\",'
-        '\\"sourceArn\\":\\"arn:aws:ses:eu-west-1:123456789012:identity/testing-notify\\",'
-        '\\"sendingAccountId\\":\\"123456789012\\",'
-        '\\"messageId\\":\\"ref1\\",'
-        '\\"destination\\":[\\"testing@digital.cabinet-office.gov.uk\\"]},'
-        '\\"delivery\\":{\\"timestamp\\":\\"2016-03-14T12:35:26.567Z\\",'
-        '\\"processingTimeMillis\\":658,'
-        '\\"recipients\\":[\\"testing@digital.cabinet-office.gov.uk\\"],'
-        '\\"smtpResponse\\":\\"250 2.0.0 OK 1457958926 uo5si26480932wjc.221 - gsmtp\\",'
-        '\\"reportingMTA\\":\\"a6-238.smtp-out.eu-west-1.amazonses.com\\"}}",'
-        '\n  "Timestamp" : "2016-03-14T12:35:26.665Z",\n  "SignatureVersion" : "1",'
-        '\n  "Signature" : "",'
-        '\n  "SigningCertURL" : "https://sns.eu-west-1.amazonaws.com/SimpleNotificationService-bb750'
-        'dd426d95ee9390147a5624348ee.pem",'
-        '\n  "UnsubscribeURL" : "https://sns.eu-west-1.amazonaws.com/?Action=Unsubscribe&S'
-        'subscriptionArn=arn:aws:sns:eu-west-1:302763885840:preview-emails:d6aad3ef-83d6-4cf3-a470-54e2e75916da"\n}'
-    )
 
 
 def sns_success_callback(reference=None, timestamp="2016-06-28 00:40:34.558", destination="+1XXX5550100"):
