+++ app/celery/research_mode_tasks.py
@@ -1,4 +1,3 @@
-import json
 import random
 from datetime import datetime, timedelta
 
@@ -6,7 +5,14 @@ from flask import current_app
 from notifications_utils.s3 import s3upload
 
 from app import notify_celery
-from app.aws.aws_mocks import sns_failed_callback, sns_success_callback
+from app.aws.aws_mocks import (
+    ses_hard_bounce_callback,
+    ses_notification_callback,
+    ses_soft_bounce_callback,
+    sns_failed_callback,
+    sns_s3_callback,
+    sns_success_callback,
+)
 from app.aws.s3 import file_exists
 from app.celery.process_ses_receipts_tasks import process_ses_results
 from app.celery.process_sns_receipts_tasks import process_sns_results
@@ -49,26 +55,6 @@ def aws_ses_callback(notification_id, to):
         return sns_success_callback("Success", destination=to, timestamp=timestamp)
 
 
-def firetext_callback(notification_id, to):
-    """
-    status: 0 - delivered
-    status: 1 - perm failure
-    status: 2 - temp failure
-    """
-    if to.strip().endswith(perm_fail):
-        status = "1"
-    elif to.strip().endswith(temp_fail):
-        status = "2"
-    else:
-        status = "0"
-    return {
-        "mobile": to,
-        "status": status,
-        "time": "2016-03-10 14:17:00",
-        "reference": notification_id,
-    }
-
-
 @notify_celery.task(
     bind=True,
     name="create-fake-letter-response-file",
@@ -106,137 +92,5 @@ def create_fake_letter_response_file(self, reference):
 
     # on development we can't trigger SNS callbacks so we need to manually hit the DVLA callback endpoint
     if current_app.config["NOTIFY_ENVIRONMENT"] == "development":
-        body = _fake_sns_s3_callback(upload_file_name)
+        body = sns_s3_callback(upload_file_name, reference)
         process_sns_results.apply_async([body], queue=QueueNames.RESEARCH_MODE)
-
-
-def _fake_sns_s3_callback(filename):
-    message_contents = '{"Records":[{"s3":{"object":{"key":"%s"}}}]}' % (filename)  # noqa
-    return json.dumps(
-        {
-            "Type": "Notification",
-            "MessageId": "some-message-id",
-            "Message": message_contents,
-        }
-    )
-
-
-def ses_notification_callback(reference):
-    ses_message_body = {
-        "delivery": {
-            "processingTimeMillis": 2003,
-            "recipients": ["success@simulator.amazonses.com"],
-            "remoteMtaIp": "123.123.123.123",
-            "reportingMTA": "a7-32.smtp-out.eu-west-1.amazonses.com",
-            "smtpResponse": "250 2.6.0 Message received",
-            "timestamp": "2017-11-17T12:14:03.646Z",
-        },
-        "mail": {
-            "commonHeaders": {
-                "from": ["TEST <TEST@notify.works>"],
-                "subject": "lambda test",
-                "to": ["success@simulator.amazonses.com"],
-            },
-            "destination": ["success@simulator.amazonses.com"],
-            "headers": [
-                {"name": "From", "value": "TEST <TEST@notify.works>"},
-                {"name": "To", "value": "success@simulator.amazonses.com"},
-                {"name": "Subject", "value": "lambda test"},
-                {"name": "MIME-Version", "value": "1.0"},
-                {
-                    "name": "Content-Type",
-                    "value": 'multipart/alternative; boundary="----=_Part_617203_1627511946.1510920841645"',
-                },
-            ],
-            "headersTruncated": False,
-            "messageId": reference,
-            "sendingAccountId": "12341234",
-            "source": '"TEST" <TEST@notify.works>',
-            "sourceArn": "arn:aws:ses:eu-west-1:12341234:identity/notify.works",
-            "sourceIp": "0.0.0.1",
-            "timestamp": "2017-11-17T12:14:01.643Z",
-        },
-        "notificationType": "Delivery",
-    }
-
-    return {
-        "Type": "Notification",
-        "MessageId": "8e83c020-1234-1234-1234-92a8ee9baa0a",
-        "TopicArn": "arn:aws:sns:eu-west-1:12341234:ses_notifications",
-        "Subject": None,
-        "Message": json.dumps(ses_message_body),
-        "Timestamp": "2017-11-17T12:14:03.710Z",
-        "SignatureVersion": "1",
-        "Signature": "[REDACTED]",
-        "SigningCertUrl": "https://sns.eu-west-1.amazonaws.com/SimpleNotificationService-[REDACTED].pem",
-        "UnsubscribeUrl": "https://sns.eu-west-1.amazonaws.com/?Action=Unsubscribe&SubscriptionArn=[REACTED]",
-        "MessageAttributes": {},
-    }
-
-
-def ses_hard_bounce_callback(reference, bounce_subtype=None):
-    return _ses_bounce_callback(reference, "Permanent", bounce_subtype)
-
-
-def ses_soft_bounce_callback(reference, bounce_subtype=None):
-    return _ses_bounce_callback(reference, "Transient", bounce_subtype)
-
-
-def _ses_bounce_callback(reference, bounce_type, bounce_subtype=None):
-    ses_message_body = {
-        "bounce": {
-            "bounceSubType": bounce_subtype or "General",
-            "bounceType": bounce_type,
-            "bouncedRecipients": [
-                {
-                    "action": "failed",
-                    "diagnosticCode": "smtp; 550 5.1.1 user unknown",
-                    "emailAddress": "bounce@simulator.amazonses.com",
-                    "status": "5.1.1",
-                }
-            ],
-            "feedbackId": "0102015fc9e676fb-12341234-1234-1234-1234-9301e86a4fa8-000000",
-            "remoteMtaIp": "123.123.123.123",
-            "reportingMTA": "dsn; a7-31.smtp-out.eu-west-1.amazonses.com",
-            "timestamp": "2017-11-17T12:14:05.131Z",
-        },
-        "mail": {
-            "commonHeaders": {
-                "from": ["TEST <TEST@notify.works>"],
-                "subject": "ses callback test",
-                "to": ["bounce@simulator.amazonses.com"],
-            },
-            "destination": ["bounce@simulator.amazonses.com"],
-            "headers": [
-                {"name": "From", "value": "TEST <TEST@notify.works>"},
-                {"name": "To", "value": "bounce@simulator.amazonses.com"},
-                {"name": "Subject", "value": "lambda test"},
-                {"name": "MIME-Version", "value": "1.0"},
-                {
-                    "name": "Content-Type",
-                    "value": 'multipart/alternative; boundary="----=_Part_596529_2039165601.1510920843367"',
-                },
-            ],
-            "headersTruncated": False,
-            "messageId": reference,
-            "sendingAccountId": "12341234",
-            "source": '"TEST" <TEST@notify.works>',
-            "sourceArn": "arn:aws:ses:eu-west-1:12341234:identity/notify.works",
-            "sourceIp": "0.0.0.1",
-            "timestamp": "2017-11-17T12:14:03.000Z",
-        },
-        "notificationType": "Bounce",
-    }
-    return {
-        "Type": "Notification",
-        "MessageId": "36e67c28-1234-1234-1234-2ea0172aa4a7",
-        "TopicArn": "arn:aws:sns:eu-west-1:12341234:ses_notifications",
-        "Subject": None,
-        "Message": json.dumps(ses_message_body),
-        "Timestamp": "2017-11-17T12:14:05.149Z",
-        "SignatureVersion": "1",
-        "Signature": "[REDACTED]",  # noqa
-        "SigningCertUrl": "https://sns.eu-west-1.amazonaws.com/SimpleNotificationService-[REDACTED]].pem",
-        "UnsubscribeUrl": "https://sns.eu-west-1.amazonaws.com/?Action=Unsubscribe&SubscriptionArn=[REDACTED]]",
-        "MessageAttributes": {},
-    }
